Add a switch for the Show all friends and important.
Have a switch for the show all friends and for only those friends whom either we owe or either they owe us.
Have this instead of search input field.

We for now dont have use for the search button as there wont be any more fields for the friends whom with we daily deal with it can be dont with the use of scroll bar.